Handle case where author is not found
